requirements.txt
# Core Dependencies
python-dotenv==1.0.0
requests==2.31.0
pandas>=2.2.0
numpy>=1.26.0

# Trading APIs
alpaca-py>=0.25.0
python-binance>=1.0.19  # Binance crypto trading API

# AI APIs
openai>=1.12.0  # Can use for DeepSeek API (OpenAI compatible)

# Data Sources
yfinance>=0.2.36
feedparser>=6.0.11  # For RSS news feeds
datasets>=2.14.0  # Hugging Face datasets for historical news
huggingface-hub>=0.20.0  # HF authentication for news datasets

# Utilities
schedule>=1.2.0
python-dateutil>=2.8.2
pytz>=2023.3

# Logging & Monitoring
colorlog>=6.8.0

# Technical Indicators
ta>=0.11.0
